+++ pydantic/main.py
@@ -97,6 +97,7 @@ class BaseConfig:
     json_encoders: Dict[AnyType, AnyCallable] = {}
     orm_mode: bool = False
     alias_generator: Optional[Callable[[str], str]] = None
+    keep_untouched: Tuple[type, ...] = ()
 
     @classmethod
     def get_field_schema(cls, name: str) -> Dict[str, str]:
@@ -169,7 +170,7 @@ def validate_custom_root_type(fields: Dict[str, Field]) -> None:
         raise TypeError('custom root type cannot allow mapping')
 
 
-TYPE_BLACKLIST = FunctionType, property, type, classmethod, staticmethod
+UNTOUCHED_TYPES = FunctionType, property, type, classmethod, staticmethod
 
 
 class MetaModel(ABCMeta):
@@ -217,10 +218,11 @@ class MetaModel(ABCMeta):
                         config=config,
                     )
 
+            untouched_types = UNTOUCHED_TYPES + config.keep_untouched
             for var_name, value in namespace.items():
                 if (
                     is_valid_field(var_name)
-                    and (annotations.get(var_name) == PyObject or not isinstance(value, TYPE_BLACKLIST))
+                    and (annotations.get(var_name) == PyObject or not isinstance(value, untouched_types))
                     and var_name not in class_vars
                 ):
                     validate_field_name(bases, var_name)