+++ HISTORY.rst
@@ -12,6 +12,7 @@ v0.31 (unreleased)
 * add advanced exclude support for ``dict``, ``json`` and ``copy``, #648 by @MrMrRobat
 * fix bug in ``GenericModel`` for models with concrete parameterized fields, #672 by @dmontagu
 * add documentation for Literal type, #651 by @dmontagu
+* add ``Config.keep_untouched`` for custom descriptors support, #679 by @MrMrRobat
 * use ``inspect.cleandoc`` internally to get model description, #657 by @tiangolo
 * add Color to schema generation, by @euri10